Ugrade Spring Boot to 2.2.2
- Update configuration property names
  - flyway.baselineOnMigrate to spring.flyway.baselineOnMigrate
  - flyway.schemas to spring.flyway.schemas
  - spring.http.multipart.maxFileSize to spring.servlet.multipart.max-file-size
  - spring.http.multipart.maxRequestSize spring.servlet.multipart.max-request-size=100MB

- Pageagle offset type int to long
- PageImpl doesnt allow null PageRequest
- PageRequest instaces created with static of-method instead of constructor
- Use SpringApplication.setWebApplicationType instead of deprecated SpringApplication.setWebEnvironment

- Upgrade dependencies required by Spring Boot 2.2.2
  - Define version for spring-security-jwt since it's not managed by Spring Boot anymore
  - Remove versions from plugins and dependencies managed by Spring Boot in 2.2.2 version
  - Update artifact and group of Spring security oauth2 dependency
  - Update group of hibernate validator dependency
  - Update artifact ID from javax.mail to jakarta.mail

- Mockito 1.x to 2.x migration
 - Remove unnecessary stubbings
 - Replace deprecated methods and classes
 - Matcher any doesnt match null anymore, fix tests expecting some value instead of null
 - Fix invalid tests that incorrectly passed with old version
 - Use instance of ApplicationProperties instead of stubbing property getters

- Other refactoring
 - Remove unused fields and variables

+++ backend/model-service/src/main/java/fi/hel/allu/model/dao/ApplicationDao.java
@@ -123,7 +123,7 @@ public class ApplicationDao {
    */
   @Transactional(readOnly = true)
   public Page<Application> findAll(Pageable pageRequest) {
-    int offset = (pageRequest == null) ? 0 : pageRequest.getOffset();
+    long offset = (pageRequest == null) ? 0 : pageRequest.getOffset();
     int count = (pageRequest == null) ? 100 : pageRequest.getPageSize();
     QueryResults<Application> queryResults = queryFactory.select(applicationBean).from(application)
         .orderBy(application.id.asc()).offset(offset).limit(count).fetchResults();
